Updated CSS & JS Injection, Changed UI behavior
CSS and JS are now injected once regardless of how many Font Awesome
fields are on the screen. Also, CSS and JS are using EE-accepted
methods for injection.

The UI now shows a list of all options when the input is in focus but
empty. This allows for simpler icon discovery. Additionally, the
selected icon now overlays the input (making it look inline with the
icon name).