Use `url_third_themes` variable
It is possible for installations to have third-party add-on themes
installed in a different location from the ExpressionEngine themes. There
are two different sets of variables for this, as referenced in the [System
Configuration Overrides][configuration]. Use the specific third-party
themes variable to reference the script.

Configuration for ExpressionEngine Themes:
* `theme_folder_url`
* `theme_folder_path`

Configuration third-party Themes:
* `url_third_themes`
* `path_third_themes`

[configuration]: https://ellislab.com/expressionengine/user-guide/general/system_configuration_overrides.html